Fix the WE cross-os build
